Update scripts to have oldIMG database option
