+++ picrust2/pipeline.py
@@ -2,22 +2,26 @@
 
 from os import path
 import sys
-from picrust2.default import (default_tables, default_pathway_map,
-                              default_ref_dir)
+from picrust2.default import (default_ref_dir_bac, default_ref_dir_arc, default_tables_bac,
+                              default_tables_arc, default_pathway_map)
 from picrust2.place_seqs import identify_ref_files
 from picrust2.util import (make_output_dir, check_files_exist, read_fasta,
-                           read_fasta_ids, system_call_check, read_seqabun)
+                           read_fasta_ids, system_call_check, read_seqabun, prune_tree)
+from picrust2.split_domains import get_lowest_nsti, combine_domain_predictions
 
 
-def full_pipeline(study_fasta,
+def full_pipeline_split(study_fasta,
                   input_table,
                   output_folder,
                   processes,
                   placement_tool,
-                  ref_dir,
+                  ref_dir1,
+                  ref_dir2,
                   in_traits,
-                  custom_trait_tables,
-                  marker_gene_table,
+                  custom_trait_tables_ref1,
+                  custom_trait_tables_ref2,
+                  marker_gene_table_ref1,
+                  marker_gene_table_ref2,
                   pathway_map,
                   rxn_func,
                   no_pathways,
@@ -30,7 +34,6 @@ def full_pipeline(study_fasta,
                   hsp_method,
                   edge_exponent,
                   min_align,
-                  skip_nsti,
                   skip_minpath,
                   no_gap_fill,
                   coverage,
@@ -40,8 +43,10 @@ def full_pipeline(study_fasta,
                   remove_intermediate,
                   verbose):
     '''Function that contains wrapper commands for full PICRUSt2 pipeline.
+    This is the version that incorporates separate placement steps for two 
+    domains (by default, bacteria and archaea).
     Descriptions of all of these input arguments/options are given in the
-    picrust2_pipeline.py script.'''
+    picrust2_pipeline_split.py script.'''
 
     # Throw warning if --per_sequence_contrib set but --stratified unset.
     if per_sequence_contrib and not stratified:
@@ -49,58 +54,98 @@ def full_pipeline(study_fasta,
               "--stratified. This means that a stratified pathway table will "
               "be output only (i.e. a stratified metagenome table will NOT "
               "be output).\n", file=sys.stderr)
+    
+    if ref_dir1 == default_ref_dir_bac:
+        out_tree_ref1 = path.join(output_folder, "bac.tre")
+        name_ref1 = 'bac'
+    else:
+        out_tree_ref1 = path.join(output_folder, "ref1.tre")
+        name_ref1 = 'ref1'
+    if ref_dir2 == default_ref_dir_arc:
+        out_tree_ref2 = path.join(output_folder, "arc.tre")
+        name_ref2 = 'arc'
+    else:
+        out_tree_ref2 = path.join(output_folder, "ref2.tre")
+        name_ref2 = 'ref2'
+    
+    # Exit if only one set of custom trait tables has been gives
+    if custom_trait_tables_ref1 is None and not custom_trait_tables_ref2 is None:
+        sys.exit("You've set some custom trait tables for reference set 1 but not "
+              "for reference set 2. Please set both of them.")
+    elif not custom_trait_tables_ref1 is None and custom_trait_tables_ref2 is None:
+        sys.exit("You've set some custom trait tables for reference set 2 but not "
+              "for reference set 1. Please set both of them.")
 
-    out_tree = path.join(output_folder, "out.tre")
-
-    if custom_trait_tables is None:
+    if custom_trait_tables_ref1 is None:
 
         # Check that specified functional categories are allowed.
-        FUNC_TRAIT_OPTIONS = ['COG', 'EC', 'KO', 'PFAM', 'TIGRFAM']
+        FUNC_TRAIT_OPTIONS = ['EC', 'KO', 'GO', 'PFAM', 'BIGG', 'CAZY', 'GENE_NAMES']
         funcs = in_traits.split(",")
         for func in funcs:
             if func not in FUNC_TRAIT_OPTIONS:
                 sys.exit("Error - specified category " + func + " is not " +
                          "one of the default categories.")
+                         
+        funcs_ref1 = funcs
+        funcs_ref2 = funcs
 
-        func_tables = default_tables
+        func_tables_ref1 = default_tables_bac
+        func_tables_ref2 = default_tables_arc
 
     else:
         # Split paths to input custom trait tables and take the basename to be
         # the function id.
-        funcs = []
-        func_tables = {}
+        funcs_ref1, funcs_ref2 = [], []
+        func_tables_ref1, func_tables_ref2 = {}, {}
 
-        for custom in custom_trait_tables.split(","):
+        for custom in custom_trait_tables_ref1.split(","):
 
             func_id = path.splitext(path.basename(custom))[0]
-            funcs.append(func_id)
-            func_tables[func_id] = custom
+            funcs_ref1.append(func_id)
+            func_tables_ref1[func_id] = custom
+            
+        for custom in custom_trait_tables_ref2.split(","):
+
+            func_id = path.splitext(path.basename(custom))[0]
+            funcs_ref2.append(func_id)
+            func_tables_ref2[func_id] = custom
 
     # Add reaction function to be in set of gene families if it is not already
     # and as long as pathways are also to be predicted.
-    if rxn_func not in funcs and not no_pathways:
+    # Note that by default this is EC, so we would only be adding it if EC wasn't 
+    # given as an option but we do want pathways to be predicted.
+    if rxn_func not in funcs_ref1 and not no_pathways:
         orig_rxn_func = rxn_func
         rxn_func = path.splitext(path.basename(rxn_func))[0]
-        funcs.append(rxn_func)
+        funcs_ref1.append(rxn_func)
+        funcs_ref2.append(rxn_func)
 
-        if rxn_func not in func_tables:
-            func_tables[rxn_func] = orig_rxn_func
+        if rxn_func not in func_tables_ref1:
+            func_tables_ref1[rxn_func] = orig_rxn_func
+        if rxn_func not in func_tables_ref2:
+            func_tables_ref2[rxn_func] = orig_rxn_func
 
-    if not skip_norm:
-        # Append marker as well, since this also needs to be run.
-        funcs.append("marker")
-        func_tables["marker"] = marker_gene_table
+    # if not skip_norm:
+    #     # Append marker as well, since this also needs to be run.
+    #     funcs_ref1.append("marker")
+    #     func_tables_ref1["marker"] = marker_gene_table
+    # Append marker
+    # funcs_ref1.append("marker")
+    # funcs_ref2.append("marker")
+    # func_tables_ref1["marker"] = marker_gene_table_ref1
+    # func_tables_ref2["marker"] = marker_gene_table_ref2
 
     # Check that all input files exist.
-    ref_msa, tree, hmm, model = identify_ref_files(ref_dir, placement_tool)
-    files2check = [study_fasta, input_table, ref_msa, tree, hmm, model] + list(func_tables.values())
+    ref_msa_ref1, tree_ref1, hmm_ref1, model_ref1 = identify_ref_files(ref_dir1, placement_tool)
+    ref_msa_ref2, tree_ref2, hmm_ref2, model_ref2 = identify_ref_files(ref_dir2, placement_tool)
+    files2check = [study_fasta, input_table, ref_msa_ref1, tree_ref1, hmm_ref1, model_ref1, ref_msa_ref2, tree_ref2, hmm_ref2, model_ref2] + list(func_tables_ref1.values())  + list(func_tables_ref2.values())
 
     if not no_pathways:
         files2check.append(pathway_map)
 
         # Throw warning if default pathway mapfile used with non-default
         # reference files.
-        if pathway_map == default_pathway_map and ref_dir != default_ref_dir:
+        if pathway_map == default_pathway_map and ref_dir1 != default_ref_dir_bac:
             print("Warning - non-default reference files specified with "
                   "default pathway mapfile of prokaryote-specific MetaCyc "
                   "pathways (--pathway_map option). This usage may be "
@@ -130,88 +175,231 @@ def full_pipeline(study_fasta,
 
     # Define folders for intermediate files (unless --remove_intermediate set).
     if remove_intermediate:
-        place_seqs_intermediate = ""
+        place_seqs_intermediate_ref1 = ""
+        place_seqs_intermediate_ref2 = ""
         pathways_intermediate = ""
     else:
         intermediate_dir = path.join(output_folder, "intermediate")
         make_output_dir(intermediate_dir)
-        place_seqs_intermediate = path.join(intermediate_dir, "place_seqs")
+        place_seqs_intermediate_ref1 = path.join(intermediate_dir, "place_seqs_"+name_ref1)
+        place_seqs_intermediate_ref2 = path.join(intermediate_dir, "place_seqs_"+name_ref2)
         pathways_intermediate = path.join(intermediate_dir, "pathways")
 
     # Run place_seqs.py.
-    place_seqs_cmd = ["place_seqs.py",
+    place_seqs_cmd_ref1 = ["place_seqs.py",
                       "--study_fasta", study_fasta,
-                      "--ref_dir", ref_dir,
-                      "--out_tree", out_tree,
+                      "--ref_dir", ref_dir1,
+                      "--out_tree", out_tree_ref1,
                       "--processes", str(processes),
-                      "--intermediate", place_seqs_intermediate,
+                      "--intermediate", place_seqs_intermediate_ref1,
                       "--min_align", str(min_align),
                       "--chunk_size", str(5000),
                       "--placement_tool", placement_tool]
+                      
+    place_seqs_cmd_ref2 = ["place_seqs.py",
+                      "--study_fasta", study_fasta,
+                      "--ref_dir", ref_dir2,
+                      "--out_tree", out_tree_ref2,
+                      "--processes", str(processes),
+                      "--intermediate", place_seqs_intermediate_ref2,
+                      "--min_align", str(min_align),
+                      "--chunk_size", str(5000),
+                      "--placement_tool", placement_tool]
+                      
 
     if verbose:
-        place_seqs_cmd.append("--verbose")
+        place_seqs_cmd_ref1.append("--verbose")
+        place_seqs_cmd_ref2.append("--verbose")
 
-    system_call_check(place_seqs_cmd, print_command=verbose,
+    system_call_check(place_seqs_cmd_ref1, print_command=verbose,
                       print_stdout=verbose, print_stderr=True)
 
     if verbose:
-        print("Finished placing sequences on output tree: " + out_tree,
+        print("Finished placing sequences on output tree for reference 1: " + out_tree_ref1,
               file=sys.stderr)
+              
+    system_call_check(place_seqs_cmd_ref2, print_command=verbose,
+                      print_stdout=verbose, print_stderr=True)
 
+    if verbose:
+        print("Finished placing sequences on output tree for reference 2: " + out_tree_ref2,
+              file=sys.stderr)
+              
     # Get predictions for all specified functions and keep track of outfiles.
-    predicted_funcs = {}
-
-    if not skip_norm:
-        # Make sure marker database is first in the list. This is because this will
-        # be run on a single core and so will be easier to identify any errors
-        # if the program exits when working on this function type.
-        funcs.insert(0, funcs.pop(funcs.index("marker")))
-
-    for func in funcs:
-        # Change output filename for NSTI and non-NSTI containing files.
-        hsp_outfile = path.join(output_folder, func + "_predicted")
-
-        if (func == "marker" and not skip_nsti) or (skip_norm and not skip_nsti):
-            hsp_outfile = hsp_outfile + "_and_nsti.tsv.gz"
-        else:
-            hsp_outfile = hsp_outfile + ".tsv.gz"
-
-        # Keep track of output filename for next step of pipeline.
-        predicted_funcs[func] = hsp_outfile
-
-        # Run hsp.py for each function database.
+    predicted_funcs_split = {}
+              
+    # Get predictions for marker gene sets for each reference
+    reference_sets = [[ref_dir1, out_tree_ref1, marker_gene_table_ref1, name_ref1], [ref_dir2, out_tree_ref2, marker_gene_table_ref2, name_ref2]]
+    for ref in reference_sets:
+        ref_dir = ref[0]
+        out_tree = ref[1]
+        marker_table = ref[2]
+        name_ref = ref[3]
+        
+        hsp_outfile = path.join(output_folder, name_ref + "_marker_predicted_and_nsti.tsv.gz")
+        
+        predicted_funcs_split[name_ref + "_marker"] = hsp_outfile
+        
+        # Run hsp.py for both reference sets
         hsp_cmd = ["hsp.py",
                    "--tree", out_tree,
                    "--output", hsp_outfile,
-                   "--observed_trait_table", func_tables[func],
+                   "--observed_trait_table", marker_table,
                    "--hsp_method", hsp_method,
                    "--edge_exponent", str(edge_exponent),
-                   "--seed", "100"]
-
-        # Add flags to command if specified.
-        if (func == "marker" and not skip_nsti) or (skip_norm and not skip_nsti):
-            hsp_cmd.append("--calculate_NSTI")
-
-        # Run marker on only 1 processor.
-        if func == "marker":
-            hsp_cmd += ["--processes", "1"]
-        else:
-            hsp_cmd += ["--processes", str(processes)]
-
+                   "--seed", "100",
+                   "--calculate_NSTI",
+                   "--processes", "1"]
+                   
         if verbose:
             hsp_cmd.append("--verbose")
 
         system_call_check(hsp_cmd, print_command=verbose,
                           print_stdout=verbose, print_stderr=True)
+                          
+    if verbose:
+        print("Finished getting marker and NSTI predictions for both domains: " +  predicted_funcs_split[name_ref1 + "_marker"] + ", " + predicted_funcs_split[name_ref2 + "_marker"] +
+              "\nNow finding the best one for each sequence.",
+              file=sys.stderr)
+    
+    # Choose which of the reference sets is best for each sequence
+    nsti_lowest_df, nsti_dom1_df, nsti_dom2_df = get_lowest_nsti(predicted_funcs_split[name_ref1 + "_marker"],
+                                                                 predicted_funcs_split[name_ref2 + "_marker"],
+                                                                 name_ref1,
+                                                                 name_ref2,
+                                                                 verbose=verbose)
+    
+    # Write the new files
+    nsti_dom1_df.to_csv(path_or_buf=path.join(output_folder, name_ref1 + "_reduced_marker_predicted_and_nsti.tsv.gz"), sep="\t")
+    nsti_dom2_df.to_csv(path_or_buf=path.join(output_folder, name_ref2 + "_reduced_marker_predicted_and_nsti.tsv.gz"), sep="\t")
+    nsti_lowest_df.to_csv(path_or_buf=path.join(output_folder, "combined_marker_predicted_and_nsti.tsv.gz"), sep="\t")
+    
+    # Check whether we still have both domains present still
+    if nsti_dom1_df.shape[0] == 0:
+        if verbose:
+            print("Don't have any " + name_ref1 + " in the study sequences. Continuing with only " + name_ref2,
+                  file=sys.stderr)
+        
+        # Update the predicted_funcs_split dictionary
+        predicted_funcs_split[name_ref1 + "_marker"] = ""
+        predicted_funcs_split[name_ref2 + "_marker"] = path.join(output_folder, name_ref2 + "_reduced_marker_predicted_and_nsti.tsv.gz")
+    elif nsti_dom2_df.shape[0] == 0:
+        if verbose:
+            print("Don't have any " + name_ref2 + " in the study sequences. Continuing with only " + name_ref1,
+                  file=sys.stderr)
+        
+        # Update the predicted_funcs_split dictionary
+        predicted_funcs_split[name_ref1 + "_marker"] = path.join(output_folder, name_ref1 + "_reduced_marker_predicted_and_nsti.tsv.gz")
+        predicted_funcs_split[name_ref2 + "_marker"] = ""
+    else:
+        # Update the predicted_funcs_split dictionary
+        predicted_funcs_split[name_ref1 + "_marker"] = path.join(output_folder, name_ref1 + "_reduced_marker_predicted_and_nsti.tsv.gz")
+        predicted_funcs_split[name_ref2 + "_marker"] = path.join(output_folder, name_ref2 + "_reduced_marker_predicted_and_nsti.tsv.gz")
+    
+    # Now prune the trees to contain only the sequences of interest for each domain
+    out_tree_ref1_red = path.join(output_folder, name_ref1 + "_reduced.tre")
+    out_tree_ref2_red = path.join(output_folder, name_ref2 + "_reduced.tre")
+    
+    ref_d1_seqs = read_fasta_ids(ref_msa_ref1)
+    ref_d2_seqs = read_fasta_ids(ref_msa_ref2)
+    
+    prune_tree(list(nsti_dom1_df.index.values)+ref_d1_seqs, out_tree_ref1, out_tree_ref1_red)
+    prune_tree(list(nsti_dom2_df.index.values)+ref_d2_seqs, out_tree_ref2, out_tree_ref2_red)
+    
+    if verbose:
+        print("Finished getting the best domain match for each sequence: " + path.join(output_folder, "combined_marker_predicted_and_nsti.tsv.gz") +
+              " Now running hsp.py for the reduced reference sets.",
+              file=sys.stderr)
+    
+    # Run hsp.py for each function database for each of the reference sets
+    reference_sets = [[name_ref1, funcs_ref1, out_tree_ref1_red, func_tables_ref1, nsti_dom1_df], [name_ref2, funcs_ref2, out_tree_ref2_red, func_tables_ref2, nsti_dom2_df]]
+    for ref in reference_sets:
+        # First get the files we'll be using
+        name = ref[0]
+        funcs = ref[1]
+        out_tree = ref[2]
+        func_tables = ref[3]
+        nsti = ref[4]
+        if nsti.shape[0] == 0: 
+            continue
+        
+        for func in funcs:
+            hsp_outfile = path.join(output_folder, name + "_" + func + "_predicted.tsv.gz")
+        
+            predicted_funcs_split[name + "_" + func] = hsp_outfile
+            
+            hsp_cmd = ["hsp.py",
+                       "--tree", out_tree,
+                       "--output", hsp_outfile,
+                       "--observed_trait_table", func_tables[func],
+                       "--hsp_method", hsp_method,
+                       "--edge_exponent", str(edge_exponent),
+                       "--seed", "100",
+                       "--processes", str(processes)]
+                       
+            if verbose:
+                hsp_cmd.append("--verbose")
+
+            system_call_check(hsp_cmd, print_command=verbose,
+                          print_stdout=verbose, print_stderr=True)
+                          
+    if verbose:
+        print("Finished getting functional predictions for all traits.",
+              file=sys.stderr)
+        
+    # Get a list of the predictions to be joined together and join them for each domain
+    # Note that this checks the names so that it is compatible with custom trait ables
+    functions_to_combine = []
+    for func in predicted_funcs_split:
+        if "marker" not in func:
+            functions_to_combine.append(func.replace(name_ref1+'_', "").replace(name_ref2+'_', ""))
+    
+    # Get only unique values
+    functions_to_combine = set(functions_to_combine)
+    
+    # Join the predictions for each domain
+    predicted_funcs = {}
+    for func in functions_to_combine:
+      
+        combining = []
+        
+        for pred in predicted_funcs_split:
+            if func in pred:
+                  combining.append(pred)
+        
+        if len(combining) == 1:
+            predicted_funcs[func] = predicted_funcs_split[combining[0]]
+            print("Warning: There was only one file for the function: "+ func + "\n"
+                  "Maybe that's fine if you used custom traits or there were no sequences "
+                  "matching one of the domains.", file=sys.stderr)
+        
+        elif len(combining) > 2:
+            sys.exit("More than two files were available for the function: " + func +"\n"
+                     "If you are using your own custom trait files, please check that you "
+                     "have not given the same table more than once for a domain.")
+        
+        else:
+            print(combining, predicted_funcs_split)
+            out_file = path.join(output_folder, "combined_" + func + "_predicted.tsv.gz")
+            combine_domain_predictions(predicted_funcs_split[combining[0]], 
+                                       predicted_funcs_split[combining[1]], 
+                                       out_file,
+                                       verbose=verbose)
+            predicted_funcs[func] = out_file
+    
+    if verbose:
+        print("Finished joining together all trait tables with the same trait name for both domains.",
+              file=sys.stderr)
+    
+    predicted_funcs["marker"] = path.join(output_folder, "combined_marker_predicted_and_nsti.tsv.gz")
 
     # Now run metagenome pipeline commands.
     # Inititalize dictionary of function names --> metagenome output files.
     func_output = {}
 
     # Loop over each function again and run metagenome pipeline.
-    for func in funcs:
-
+    for func in predicted_funcs:
+      
         if func == "marker":
             continue
 
@@ -238,8 +426,7 @@ def full_pipeline(study_fasta,
         if wide_table:
             metagenome_pipeline_cmd.append("--wide_table")
 
-        if not skip_nsti:
-            metagenome_pipeline_cmd += ["--max_nsti", str(max_nsti)]
+        metagenome_pipeline_cmd += ["--max_nsti", str(max_nsti)]
 
         if skip_norm:
             metagenome_pipeline_cmd.append("--skip_norm")
