+++ scripts/add_descriptions_oldIMG.py
@@ -2,7 +2,7 @@
 
 import argparse
 from importlib.metadata import version
-from picrust2.default import default_map
+from picrust2.default_oldIMG import default_map
 from picrust2.util import add_descrip_col, make_output_dir_for_file
 import sys
